backend/requirements.txt
asgiref==3.5.2
autopep8==1.6.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Django==4.0.5
django-cors-headers==3.13.0
django-environ==0.9.0
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
drf-yasg==1.21.3
environ==1.0
fonttools==4.33.3
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.10.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy==1.22.3
oauthlib==3.2.0
packaging==21.3
pandas==1.4.3
phonenumbers==8.12.53
Pillow==9.1.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-surprise==1.1.1
scipy==1.9.0
six==1.16.0
sqlparse==0.4.2
surprise==0.1
toml==0.10.2
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.10
